Added PluginIntegration interface to allow for other interfaces, that are independent of plugin dependencies.
